* Fixed a bunch of compiler warnings and did some code tabbing cleanup. * Changed a lot of variables to time_t as time_t can be a 64-bit variable while an int can be a 32-bit variable depending on the platform. * Fixed NPC shield.  I somehow failed to fully commit my fix for the past two revisions. * Finally fixed all the NPC props.  Shed a tear at how I did it.  It is very sad and hacky. (see CPlayer::msgNPCPROPS)

+++ GServer217/trunk/src/CBuffer.cpp
@@ -161,7 +161,7 @@ bool CBuffer::load(const char* pFileName)
 		return false;
 	clear();
 	int size = 0;
-	while((size = fread(buffer, 1, sizeof(buffer), fHandle)) > 0)
+	while((size = (int)fread(buffer, 1, sizeof(buffer), fHandle)) > 0)
 		writeBytes(buffer, size);
 	fclose(fHandle);
 	return true;
@@ -678,9 +678,9 @@ CBuffer& CBuffer::untokenize()
 
 		// Move forward the correct number of spaces.
 		if ( pos[0] + 1 != nData.length() && nData[pos[0] + 1] == '\"' )
-			pos[1] = pos[0] + strlen( ",\"" );	// test,"test
+			pos[1] = pos[0] + (int)strlen( ",\"" );	// test,"test
 		else
-			pos[1] = pos[0] + strlen( "," );	// test,test
+			pos[1] = pos[0] + (int)strlen( "," );	// test,test
 	}
 
 	// Try and grab the very last element.
